+++ dockerGrader/initEnv.sh
@@ -1,1 +1,2 @@
+sudo apt-get install rabbitmq-server
 python3 -m venv venv && . venv/bin/activate && pip install -r requirements.txt && deactivate && echo "Success!"
