+++ dockerGrader/requirements.txt
@@ -1,14 +1,24 @@
+amqp==2.3.2
+Babel==2.6.0
+billiard==3.5.0.4
 boto3==1.9.0
 botocore==1.12.0
+celery==4.2.1
 click==6.7
 docutils==0.14
 Flask==1.0.2
+flower==0.9.2
 itsdangerous==0.24
 Jinja2==2.10
 jmespath==0.9.3
+kombu==4.2.1
 MarkupSafe==1.0
 python-dateutil==2.7.3
+pytz==2018.5
+redis==2.10.6
 s3transfer==0.1.13
 six==1.11.0
+tornado==5.1.1
 urllib3==1.23
+vine==1.1.4
 Werkzeug==0.14.1
