+++ dockerGrader/run-celery.sh
@@ -1,0 +1,1 @@
+celery -A testServer.celery worker --concurrency 2 -E
