+++ dockerGrader/run-redis.sh
@@ -1,0 +1,1 @@
+redis-server ./redis.conf
